agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "okk-agent"
version = "0.1.0"
description = "AI-driven continuous verification agent for Oxia"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "openai>=1.50.0",
    "kubernetes>=31.0.0",
    "PyGithub>=2.5.0",
    "prometheus-api-client>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
okk-agent = "okk_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/okk_agent"]

[tool.ruff]
line-length = 120
